Renamed some things; updated logging to be more specific.
